Move Fuse to `cs` folder
The manifest is no longer modified by this PR.
