+++ manifest.json
@@ -56,8 +56,7 @@
     "addons/*",
     "libraries/*/cs/*",
     "addons-l10n/*/*.json",
-    "images/cs/*",
-    "libraries/thirdparty/*"
+    "images/cs/*"
   ],
   "browser_specific_settings": {
     "gecko": {
