+++ webpages/settings/index.html
@@ -45,7 +45,7 @@
     <link rel="modulepreload" href="../../libraries/common/cs/download-blob.js" />
     <link rel="modulepreload" href="../rtl-list.js" />
     <link rel="modulepreload" href="../../libraries/common/load-vue-components.js" />
-    <link rel="modulepreload" href="../../libraries/thirdparty/fuse.esm.min.js" />
+    <link rel="modulepreload" href="../../libraries/thirdparty/cs/fuse.esm.min.js" />
     <link rel="modulepreload" href="data/tags.js" />
     <link rel="modulepreload" href="data/addon-groups.js" />
     <link rel="modulepreload" href="data/categories.js" />
